Added a check to ensure assets actually exist before copying.

+++ src/Deployer.php
@@ -368,7 +368,13 @@ class Deployer
         $commands = [];
 
         foreach ($assets as $asset => $target) {
+            if (! file_exists($asset)) {
+                $commands[] = 'echo '.$asset . ' not found. Skipping.';
+                continue;
+            }
+
             $target = Str::startsWith($target, '/') ? $target : '/'.$target;
+
 
             $commands[] = 'scp -P'
                 .$this->server->port() // port
